+++ photon-model-adapters/aws/src/main/java/com/vmware/photon/controller/model/adapters/awsadapter/AWSConstants.java
@@ -18,7 +18,6 @@ import com.vmware.photon.controller.model.UriPaths;
 public class AWSConstants {
     public static final String AWS_SECURITY_GROUP = "awsSecurityGroup";
     public static final String AWS_SECURITY_GROUP_ID = "awsSecurityGroupId";
-    public static final String AWS_INSTANCE_ID = "awsInstanceId";
     public static final String AWS_VPC_ID = "awsVpcId";
     public static final String AWS_SUBNET_ID = "awsSubnetId";
     public static final String INSTANCE_STATE = "instance-state-name";